Move RsModuleType and friends into core module
And finally drop hack from `plugin.xml` to register rust module
