+++ idea/src/main/resources/META-INF/idea-only.xml
@@ -1,6 +1,5 @@
 <idea-plugin>
     <extensions defaultExtensionNs="com.intellij">
-        <moduleBuilder builderClass="org.rust.ide.idea.RsModuleBuilder"/>
         <projectStructureDetector implementation="org.rust.ide.idea.RsProjectStructureDetector"/>
     </extensions>
 </idea-plugin>
