+++ plugin/src/main/resources/META-INF/plugin.xml
@@ -29,11 +29,4 @@
     <depends optional="true" config-file="grazie-only.xml">tanvd.grazi</depends>
     <depends optional="true" config-file="js-only.xml">JavaScript</depends>
     <depends optional="true" config-file="ml-completion-only.xml">com.intellij.completion.ml.ranking</depends>
-
-    <extensions defaultExtensionNs="com.intellij">
-        <!-- although Rust module type is only created by IDEA, we need it in other IDEs as well
-             to be able to open a project created in IDEA without errors -->
-        <moduleType id="RUST_MODULE" implementationClass="org.rust.ide.idea.RsModuleType"/>
-    </extensions>
-
 </idea-plugin>
