+++ src/main/kotlin/org/rust/ide/module/RsModuleBuilder.kt
@@ -3,7 +3,7 @@
  * found in the LICENSE file.
  */
 
-package org.rust.ide.idea
+package org.rust.ide.module
 
 import com.intellij.execution.ExecutionException
 import com.intellij.ide.util.projectWizard.ModuleBuilder
